+++ src/components/common/Expenditures/ExpenditureForm/ExpenditureForm.module.css
@@ -23,6 +23,7 @@
   display: flex;
   align-items: center;
   justify-content: flex-end;
+  margin-top: 10px;
 }
 
 .cancelButton {
