+++ src/components/common/Expenditures/ExpenditureForm/index.ts
@@ -1,3 +1,3 @@
-export { default } from './ExpenditureForm';
-export * from './ExpenditureForm';
+export { default as EditExpenditureForm } from './EditExpenditureForm';
+export { default as CreateExpenditureForm } from './CreateExpenditureForm';
 export * from './types';
