+++ src/components/common/Expenditures/ExpenditureForm/types.ts
@@ -1,3 +1,7 @@
+export interface ExpenditureFormValues {
+  payouts: ExpenditurePayoutFieldValue[];
+}
+
 export interface ExpenditurePayoutFieldValue {
   slotId?: number;
   recipientAddress: string;
