+++ src/components/common/Expenditures/ExpendituresPage/ExpendituresPage.tsx
@@ -6,7 +6,7 @@ import { useColonyContext } from '~hooks';
 import { notNull } from '~utils/arrays';
 import Link from '~shared/Link';
 
-import ExpenditureForm from '../ExpenditureForm';
+import { CreateExpenditureForm } from '../ExpenditureForm';
 
 import styles from './ExpendituresPage.module.css';
 
@@ -43,7 +43,7 @@ const ExpendituresPage = () => {
 
       <div>
         <Heading3>Create new expenditure</Heading3>
-        <ExpenditureForm />
+        <CreateExpenditureForm />
       </div>
     </div>
   );
